🧪✨ added e2e testing workflow
