+++ .github/workflows/main.yml
@@ -34,8 +34,15 @@ jobs:
         run: blender --version
       
       # Runs a single command using the runners shell
-      - name: List Content Of Current Dir
-        run: ls -la
+      - name: Run E2E Tests
+        run: blender --background -noaudio --python test/e2e.py
+
+      # preserve artifacts for later investigation
+      - name: Preserve Artifacts For Later Investigation
+        uses: actions/upload-artifact@v2
+        with:
+          name: e2e-testing-result
+          path: test/dist/
 
       # Runs a set of commands using the runners shell
       # - name: Run a multi-line script
